Add files via upload
This web application is a parcel tracking system that was built on my university server and apart form the web languages used previously i used php and mysql. So this one is a 3 user system each with a different functionality added unto them. you begin with the visitor page where you can then login as either a registered user or admin